spliting functions into ft_linkedlist.c file and ft_swap.c file
