+++ program_to_test/Makefile
@@ -6,7 +6,7 @@
 #    By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/05/09 17:59:28 by umeneses          #+#    #+#              #
-#    Updated: 2024/05/20 15:38:56 by umeneses         ###   ########.fr        #
+#    Updated: 2024/05/22 11:47:08 by umeneses         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -45,7 +45,8 @@ LIBS					= $(LIBFT)
 NAME					= push_swap
 
 SRC_FILES				= push_main.c
-SRC_FILES				+= push_functions.c
+SRC_FILES				+= ft_linkedlist.c
+SRC_FILES				+= ft_swap.c
 
 SRC_FILES_ALL			= $(addprefix $(SRC_D), $(SRC_FILES))
